Splitted dialog in more components
