--- src/components/dialogs/ExportDialog.css
@@ -1,71 +1,0 @@
-.export-dialog h3 {
-  margin-left: 1rem;
-  margin-right: 1rem;
-}
-
-.export-dialog hr {
-  border: none;
-  height: 2px;
-  background-color: black;
-  margin-bottom: 1rem;
-}
-
-.export-dialog .row {
-  display: flex;
-  margin-bottom: 1rem;
-  align-items: center;
-}
-
-.export-dialog .row>div {
-  margin-left: 1rem;
-}
-
-.export-dialog input {
-  line-height: calc(2rem + 2px);
-}
-
-.export-dialog a {
-  display: none;
-}
-
-.export-dialog .button {
-  margin-right: 1rem;
-  border: 2px solid black;
-  line-height: 2rem;
-  padding-left: 1rem;
-  padding-right: 1rem;
-  background-color: #3d7deb;
-  color: white;
-}
-
-.export-dialog .button:hover {
-  background-color: #2e63bf;
-}
-
-.export-dialog .button:active {
-  background-color: #234a8d;
-}
-
-.export-dialog .button.close {
-  background-color: transparent;
-  color: black;
-  border: none;
-  float: right;
-  margin: 0;
-  padding: 0;
-  margin-left: 1rem;
-}
-
-.export-dialog .button.close:hover {
-  color: #3a3a3a;
-}
-
-.export-dialog .button.close:active {
-  color: #7e7e7e;
-}
-
-.export-dialog .message {
-  width: 100%;
-  margin-right: 1rem;
-  text-align: center;
-}
