+++ src/components/dialogs/ExportDialogBody.tsx
@@ -2,7 +2,7 @@ import React, { useRef, useState } from "react";
 import { hot } from "react-hot-loader";
 import DatePicker from "react-datepicker";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faCheck, faXmark } from "@fortawesome/free-solid-svg-icons";
+import { faCheck } from "@fortawesome/free-solid-svg-icons";
 
 import { useAppDispatch, useCurrentDate } from "../../redux/hooks";
 import * as ConnectionErrorSlice from "../../redux/connectionErrorSlice";
@@ -11,27 +11,20 @@ import * as Utils from "../../utils";
 import * as Api from "../../api";
 
 import "react-datepicker/dist/react-datepicker.css";
-import "./ExportDialog.css";
 
 type DialogState = "fill" | "loading" | "done" | "no data";
 
 type Props = {
   type: DialogSlice.DialogType,
-  dialogRef: React.RefObject<HTMLDivElement>,
-  fadeOutDialog: () => void,
-  onAnimationEnd: () => void
+  fadeOutDialog: () => void
 }
 
-function ExportDialog(props: Props): JSX.Element {
+function ExportDialogBody(props: Props): JSX.Element {
   const dispatch = useAppDispatch();
   const currentDate = useCurrentDate();
   const [selectedDate, setSelectedDate] = useState(currentDate);
   const [dialogState, setDialogState] = useState<DialogState>("fill");
   const anchorRef = useRef<HTMLAnchorElement>(null);
-
-  function preventHideOnSelfClick(event: React.MouseEvent<HTMLDivElement>) {
-    event.stopPropagation();
-  }
 
   function onDateChange(date: Date) {
     setSelectedDate(Utils.dateToString(date));
@@ -39,7 +32,6 @@ function ExportDialog(props: Props): JSX.Element {
 
   let getDataAsync: () => Promise<Blob>;
   let filename: string;
-  let title: string;
 
   switch (props.type) {
   case "police":
@@ -48,7 +40,6 @@ function ExportDialog(props: Props): JSX.Element {
       return response.data;
     };
     filename = `polizia-${selectedDate}.txt`;
-    title = "Esporta Dati Polizia";
     break;
   case "istat":
     getDataAsync = async () => {
@@ -56,7 +47,6 @@ function ExportDialog(props: Props): JSX.Element {
       return response.data;
     };
     filename = `istat-${selectedDate}.pdf`;
-    title = "Esporta Dati ISTAT";
     break;
   }
 
@@ -86,10 +76,10 @@ function ExportDialog(props: Props): JSX.Element {
     setDialogState("loading");
   }
 
-  let dialogBody = <></>;
+  let dialogRow = <></>;
   switch (dialogState) {
   case "fill":
-    dialogBody = (
+    dialogRow = (
       <>
         <div>
           <DatePicker
@@ -104,32 +94,22 @@ function ExportDialog(props: Props): JSX.Element {
     );
     break;
   case "loading":
-    dialogBody = (<div className="message">Esporto...</div>);
+    dialogRow = (<div className="message">Esporto...</div>);
     break;
   case "done":
-    dialogBody = (<div className="message">Fatto <FontAwesomeIcon icon={faCheck} /></div>);
+    dialogRow = (<div className="message">Fatto <FontAwesomeIcon icon={faCheck} /></div>);
     break;
   case "no data":
-    dialogBody = (<div className="message">Nessun dato da esportare!</div>);
+    dialogRow = (<div className="message">Nessun dato da esportare!</div>);
     break;
   }
 
   return (
-    <div
-      ref={props.dialogRef}
-      className="export-dialog show"
-      onClick={preventHideOnSelfClick}
-      onAnimationEnd={props.onAnimationEnd}
-    >
-      <h3>
-        {title}
-        <FontAwesomeIcon className="button close" icon={faXmark} onClick={props.fadeOutDialog} />
-      </h3>
-      <hr />
-      <div className="row">{dialogBody}</div>
+    <>
+      <div className="row">{dialogRow}</div>
       <a ref={anchorRef}></a>
-    </div>
+    </>
   );
 }
 
-export default hot(module)(ExportDialog);
+export default hot(module)(ExportDialogBody);
